third_party/blink/web_tests/wpt_internal/battery-status/multiple-promises-after-resolve.html

<!DOCTYPE html>
<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
import {MockBatteryMonitor} from './resources/mock-battery-monitor.js';

const mock = new MockBatteryMonitor();

promise_test(async _ => {
  const promise1 = navigator.getBattery();
  mock.setBatteryStatus(false, 10, 20, 0.5);
  const result1 = await promise1;
  mock.verifyBatteryStatus(result1);
  const promise2 = navigator.getBattery();
  assert_equals(promise1, promise2);
  const result2 = await promise2;
  mock.verifyBatteryStatus(result2);
}, 'multiple consecutive invocations of navigator.getBattery()');
</script>
</body>
</html>
